Slight cleanup and keeping around bootstrapper

+++ src/EZUtils.PackageManager/BootstrapPackage/BootstrapperPass2.cs
@@ -14,6 +14,7 @@ namespace EZUtils.PackageManager
         public static async void Run()
         {
             if (File.Exists("Assets/EZUtils/BootstrapPackage/Block.txt")) return;
+            if (File.Exists($"Assets/EZUtils/BootstrapPackage/{TargetPackageName}/Finished.txt")) return;
 
             PackageRepository repo = new PackageRepository();
             repo.CheckForScopedRegistry();
@@ -23,8 +24,12 @@ namespace EZUtils.PackageManager
             PackageInformation targetPackage = packages.SingleOrDefault(p => p.Name == TargetPackageName);
             if (targetPackage == null) throw new InvalidOperationException(
                 $"Could not install package '{TargetPackageName}'.");
-
             await repo.SetVersionAsync(targetPackage, targetPackage.Versions[0]);
+
+            PackageInformation packageManagerPackage = packages.SingleOrDefault(p => p.Name == "com.timiz0r.ezutils.packagemanager");
+            if (packageManagerPackage == null) throw new InvalidOperationException(
+                $"Could not install package manager.");
+            await repo.SetVersionAsync(packageManagerPackage, targetPackage.Versions[0]);
 
             _ = AssetDatabase.DeleteAsset($"Assets/EZUtils/BootstrapPackage/{TargetPackageName}");
         }
